Convert to slash commands and fix file deletion
The server no longer uses the Slack events API to listen for file
shares and respond to them. Instead, it can now be invoked via a slash
command. The slash command sends a file URL (either from Slack or an
external source) to /view on the server.

In addition, the server should now be able to handle deletions more
reliably. Before, it set a timeout for the deletion. However, since the
server is put to sleep after 30 minutes of inactivity on Heroku, it
seemed to cause problems with the deletion. Now, the server instead does
a "sweep" of the molecules directory after a few uploads and deletes old
files by looking at the timestamp in each file's name.