Refactor to use OAuth
This commit brings vast changes that allow MOVES to use Slack's OAuth in
a much more flexible manner. In the past, MOVES relied on a hard-coded
OAuth token for the Mobley Lab workspace (stored in an environment
variable, of course). Now, it is able to properly interface with Slack
OAuth. This is accomplished through several processes.

First of all, MOVES stores OAuth tokens in a dictionary, with a Slack
team's team ID as the key. When MOVES is closed, it writes the team IDs
and tokens to a local file. When MOVES starts again, it reads from this
file.

Second of all, MOVES now implements a redirect URL to allow it to obtain
OAuth tokens from Slack using its client ID and client secret. This
means that MOVES can now be installed to any workspace. To make such
installation easier, a webpage with an add_to_slack button is mounted at
/auth; the page itself is stored in add_to_slack.html.