Move to AWS for file storage
With this commit, S3, a part of AWS (Amazon Web Services) is now used
for storing the app's files. This means that files can finally be stored
permanently, instead of disappearing whenever the server on Heroku dies.
Currently, the files being stored are the CSV file with the team ID's
and OAuth tokens, and the molecule files downloaded from Slack. The
molecule files are public, and the CSV file is, of course, private.

This commit is being made without much attention to style in the code. A
future commit will address refactoring and other cleaning up.