Configure eslint and prettier to work together
This commit configures eslint and prettier to work together for linting
and formatting files. For now, the Google style is used for linting with
eslint, while prettier handles formatting. Using husky and lint-staged,
a precommit hook was added to run eslint, prettier, and markdown-toc on
relevant files whenever a commit is attempted (see package.json for
these configurations).

See these two tutorials on prettier and eslint by 39digits for more
information:
- https://www.39digits.com/configure-prettier-and-eslint-in-visual-studio-code/
- https://www.39digits.com/automatically-format-your-javascript-commits-using-prettier-and-husky/